+++ NEWS.md
@@ -3,9 +3,10 @@ CohortGenerator 0.2.0
 
 - Updated documentation
 - Remove CirceR dependency and related functions
-- Refactoring many functions to enable easier chainging of function call with 
-ROhdsiWebApi and CohortDiagnostics
+- Refactoring many functions to enable easier chaining of function calls with ROhdsiWebApi and CohortDiagnostics
 - Add getCohortCounts function
+- Add exportCohortStatsTables function for exporting cohort statistics to the file system 
+- Add dropCohortStatsTables function for dropping cohort statistics tables
 
 CohortGenerator 0.1.1
 =======================